+++ models/Translation.js
@@ -15,10 +15,11 @@ var Translation = new keystone.List('Translation', {
 });
 
 Translation.add({
+  title: { type: Types.Text, collapse: true },
   when: { type: Types.Date, default: Date.now, initial: true },
   author: { type: Types.Relationship, ref: 'Student', hidden: true  },
   authors: { type: Types.Relationship, ref: 'Student', initial: true, index: true, many: true },
-  editor: { type: Types.Relationship, ref: 'Student', filters: { canEditArticles: true }}
+  editor: { type: Types.Relationship, ref: 'Student', filters: { canEditArticles: true }, collapse: true }
 }, 'Meta', {
   partial: { type: Types.Boolean, default: false, hidden: true },
   multipleAuthors: { type: Types.Boolean, noedit: true },
@@ -113,6 +114,27 @@ Translation.schema.post('remove', function() {
 		if (student) student.refreshTranslations();
 	});
 })
+
+
+Translation.schema.methods.toCSV = function(row, callback) {
+
+  // we don't need the previously used fields
+  delete row.author;
+  delete row.partial;
+
+  callback(null, row);
+
+  /*if (rtn.editor != undefined) {
+    rtn.editor_id = rtn.editor.key;
+  }
+
+  this.populate('authors', function() {
+      rtn.author_ids = translation.authors.map(function(i) { return i.key });
+      callback(null, rtn);
+  });*/
+
+};
+
 
 /**
  * Registration